requirements.txt
langchain
langchain-community
langsmith

groq

transformers
torch

pymysql
sqlalchemy
mysql-connector-python
cryptography

python-dotenv

pandas